backend/requirements.txt
# API
fastapi
uvicorn[standard]

# LLM & RAG
openai
python-dotenv
pymilvus
sentence-transformers
PyMuPDF